changed ID value from Long to Integer

+++ src/main/java/spring/framework/petclinic/services/OwnerService.java
@@ -2,6 +2,6 @@ package spring.framework.petclinic.services;
 
 import spring.framework.petclinic.model.Owner;
 
-public interface OwnerService extends CrudService<Owner, Long> {
+public interface OwnerService extends CrudService<Owner, Integer> {
     Owner findByLastName(String lastName);
     }
