+++ src/main/java/spring/framework/petclinic/services/PetService.java
@@ -2,6 +2,6 @@ package spring.framework.petclinic.services;
 
 import spring.framework.petclinic.model.Pet;
 
-public interface PetService extends CrudService<Pet, Long> {
+public interface PetService extends CrudService<Pet, Integer> {
 
 }
