+++ src/main/java/spring/framework/petclinic/services/VetService.java
@@ -2,6 +2,6 @@ package spring.framework.petclinic.services;
 
 import spring.framework.petclinic.model.Vet;
 
-public interface VetService extends CrudService<Vet, Long> {
+public interface VetService extends CrudService<Vet, Integer> {
 
 }
